Wrap the tokio runtime in a Mutex.
After quite a bit of debate before, we decided to wrap the tokio executor in an AtomicRefCell.
Thanks to help from the tokio team, we've been able to figure out why tokios spawn, spawn_blocking and block_on take a mutable reference to the runtime, trigger a panic using AtomicRefCell, and make sure it doesn't happen again thanks to a mutex.
We also added a test to make sure it won't ever happen anymore.
Thanks again to some tokio contributors who helped us figure out what is going on and how to trigger the error.